refactor: dto를 engine 패키지로 옮김
