+++ src/main/java/me/namuhuchutong/ladder/domain/engine/dto/LadderResult.java
@@ -1,4 +1,4 @@
-package me.namuhuchutong.ladder.dto;
+package me.namuhuchutong.ladder.domain.engine.dto;
 
 import me.namuhuchutong.ladder.domain.implement.Names;
 import me.namuhuchutong.ladder.domain.implement.Results;
