+++ src/test/java/me/namuhuchutong/ladder/UserInputInformationTest.java
@@ -1,6 +1,6 @@
 package me.namuhuchutong.ladder;
 
-import me.namuhuchutong.ladder.dto.UserInputInformation;
+import me.namuhuchutong.ladder.domain.engine.dto.UserInputInformation;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
